(chore) folders restructuration
