+++ projects/lib/src/configurator/public-api.ts
@@ -1,7 +1,7 @@
 export * from './palette/palette.component';
 export * from './editors';
 export * from './tree/tree.component';
-export * from './toolbar.component';
+export * from './toolbar/toolbar.component';
 export * from './configurator.component';
 export * from './configurator.module';
 export * from './configurator.models';
