+++ projects/lib/src/configurator/toolbar/toolbar.component.html
