+++ projects/lib/src/dynamic-views/dynamic-views.module.ts
@@ -1,12 +1,13 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
+import { DndModule } from 'ngx-drag-drop';
+
 import { ConfigurableModule } from '../configurable/configurable.module';
 import { UtilsModule } from '../utils/utils.module';
-import { DndModule } from 'ngx-drag-drop';
 import { ConfigModule } from '../configuration';
 
-import { ZoneComponent } from './zone.component';
-import { ItemComponent } from './item.component';
+import { ZoneComponent } from './zone/zone.component';
+import { ItemComponent } from './item/item.component';
 
 @NgModule({
   imports: [CommonModule, DndModule, ConfigurableModule, UtilsModule, ConfigModule],
