+++ projects/lib/src/dynamic-views/item/item.component.ts
@@ -12,10 +12,10 @@ import {
 } from '@angular/core';
 import { DndDropEvent } from 'ngx-drag-drop';
 import { Subscription } from 'rxjs';
-import { ConditionsService } from '../conditions/conditions.service';
-import { ComponentConfig, ConfigService } from '../configuration';
-import { TemplateNameDirective } from '../utils';
-import { ContainerIndex, DragDropService } from './drag-drop.service';
+import { ConditionsService } from '../../conditions/conditions.service';
+import { ComponentConfig, ConfigService } from '../../configuration';
+import { TemplateNameDirective } from '../../utils';
+import { ContainerIndex, DragDropService } from '../drag-drop.service';
 
 @Component({
   selector: '[uib-item]',
