+++ projects/lib/src/dynamic-views/public-api.ts
@@ -1,4 +1,4 @@
 export * from './drag-drop.service';
 export * from './dynamic-views.module';
-export * from './item.component';
-export * from './zone.component';
+export * from './item/item.component';
+export * from './zone/zone.component';
