+++ projects/lib/src/utils/public-api.ts
@@ -1,8 +1,8 @@
-export * from './autocomplete.component';
-export * from './template-name.directive';
-export * from './modal.component';
-export * from './tooltip.directive';
+export * from './autocomplete/autocomplete.component';
+export * from './directive/template-name.directive';
+export * from './modal/modal.component';
+export * from './directive/tooltip.directive';
 export * from './utils.module';
-export * from './model-change.directive';
+export * from './directive/model-change.directive';
 export * from './toast/toast.component'
 export * from './toast/toast.service';
