+++ projects/lib/src/utils/utils.module.ts
@@ -1,10 +1,10 @@
 import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
-import { TooltipDirective } from './tooltip.directive';
-import { ModalComponent } from './modal.component';
-import { TemplateNameDirective } from './template-name.directive';
-import { AutocompleteComponent } from './autocomplete.component';
-import { NgModelChangeDebouncedDirective } from './model-change.directive';
+import { TooltipDirective } from './directive/tooltip.directive';
+import { ModalComponent } from './modal/modal.component';
+import { TemplateNameDirective } from './directive/template-name.directive';
+import { AutocompleteComponent } from './autocomplete/autocomplete.component';
+import { NgModelChangeDebouncedDirective } from './directive/model-change.directive';
 import { ToastComponent } from './toast/toast.component';
 
 @NgModule({
